Replace GTK clipboard with copypasta
GTK clipboard showed rather inconsistent behavior in my testing of copying on Windows platform, crashing my programs in the process. The reason of this behavior is unknown, but since copypasta works as expected and doesn't require operation on the GTK thread (meaning cleaner code), I decided to go prefer it over the previous approach.